Test.Framework.Seed

Documentation

data Seed

Constructors

FixedSeed Int

RandomSeed

newSeededStdGen :: Seed -> IO (StdGen, Int)

Given a Seed, returns a new random number generator based on that seed and the actual numeric seed that was used to build that generator, so it can be recreated.

newStdGenWithKnownSeed :: IO (StdGen, Int)