Test.Framework.Options

Documentation

type TestOptions = TestOptions' Maybe

type CompleteTestOptions = TestOptions' K

data TestOptions' f

Constructors

TestOptions

Fields

topt_seed :: f Seed

Seed that should be used to create random numbers for generated tests

topt_maximum_generated_tests :: f Int

Maximum number of tests to generate when using something like QuickCheck

topt_maximum_unsuitable_generated_tests :: f Int

Maximum number of unsuitable tests to consider before giving up when using something like QuickCheck

topt_maximum_test_size :: f Int

Maximum size of generated tests when using something like QuickCheck

topt_maximum_test_depth :: f Int

Maximum depth of generated tests when using something like SmallCheck

topt_timeout :: f (Maybe Int)

The number of microseconds to run tests for before considering them a failure